Init queue dao

+++ dao/QueueDao.py
@@ -9,13 +9,13 @@ from dao import set_default
 table_name = "QueueTable"
 
 class QueueRecord:
-  def __init__(self, guild_id: str, queue_id: str, team_1: set, team_2: set, queue: set, version: float):
+  def __init__(self, guild_id: str, queue_id: str, team_1: set, team_2: set, queue: set, version: int):
     self.guild_id = guild_id
     self.queue_id = queue_id
     self.team_1 = team_1
     self.team_2 = team_2
     self.queue = queue
-    self.version = float(version)
+    self.version = int(version)
 
   def clear_queue(self):
     self.team_1 = set()
